[btree] 加入btree helper模块，代码结构更清晰
